requirements.txt
azure-functions
openai>=1.30.0
chromadb>=0.5.0
requests>=2.31.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
